requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
ase==3.22.1
beautifulsoup4==4.12.3
Bio==1.7.1
biopython==1.84
chatmof==0.2.0
dashscope==1.20.0
emmet-core==0.84.2rc4
fastapi==0.115.5
feedparser==6.0.11
generate-chroma==1.0.0
html2text==2020.1.16
httpx==0.27.2
ipython==8.12.3
keras==2.15.0
langchain==0.3.8
langchain-community==0.3.8
langchain-experimental==0.3.3
llama-index==0.11.3
matplotlib==3.9.2
MDAnalysis==2.8.0
moffragmentor==0.0.6
molbloom==2.1.0
molSimplify==1.7.5
mp-api==0.41.2
networkx==3.2.1
nltk==3.8.1
numpy==2.1.3
openai==1.55.0
pandas==2.2.3
paper-qa==1.1.1
Pillow==11.0.0
ProDy==2.4.1
pymatgen==2023.12.18
python-dotenv==1.0.1
rdkit==2024.3.3
Requests==2.32.3
RXN4Chemistry==1.13.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selfies==2.1.1
tiktoken==0.5.2
torch==1.12.0+cu113
tqdm==4.66.1
transformers==4.41.2
uvicorn==0.32.1
